+++ source/DSCResources/AddsOrgUnitsAndGroups/AddsOrgUnitsAndGroups.schema.psm1
@@ -103,8 +103,17 @@ configuration AddsOrgUnitsAndGroups
         if ($group.GroupScope -eq 'DomainLocal')
         {
             $dependencies += "[ADGroup]'$($group.GroupName)'"
-            $group.DependsOn = $ouDependencies
             $group.Path = '{0},{1}' -f $group.Path, $DomainDn
+
+            $parentOuRef = "[ADOrganizationalUnit]$($group.Path -replace '\W')"
+            if ($ouDependencies -contains $parentOuRef)
+            {
+                $group.DependsOn = $parentOuRef
+            }
+            else
+            {
+                $group.DependsOn = $ouDependencies
+            }
         }
         elseif (($group.GroupScope -eq 'Global') -or (-not [string]::IsNullOrWhiteSpace($group.Path)))
         {